+++ src/main/java/net/frozenblock/wilderwild/registry/WWCriteria.java
@@ -19,6 +19,7 @@
 package net.frozenblock.wilderwild.registry;
 
 import net.frozenblock.wilderwild.WWConstants;
+import net.frozenblock.wilderwild.advancement.FragileIceFallOntoAndBreakTrigger;
 import net.frozenblock.wilderwild.advancement.MobBottleTrigger;
 import net.frozenblock.wilderwild.advancement.TermiteEatTrigger;
 import net.minecraft.advancements.CriteriaTriggers;
@@ -26,6 +27,7 @@ import net.minecraft.advancements.CriteriaTriggers;
 public final class WWCriteria {
 	public static final MobBottleTrigger MOB_BOTTLE = CriteriaTriggers.register(WWConstants.string("mob_bottle"), new MobBottleTrigger());
 	public static final TermiteEatTrigger TERMITE_EAT = CriteriaTriggers.register(WWConstants.string("termite_eat"), new TermiteEatTrigger());
+	public static final FragileIceFallOntoAndBreakTrigger FRAGILE_ICE_FAL_ONTO_AND_BREAK = CriteriaTriggers.register(WWConstants.string("fragile_ice_fall_onto_and_break"), new FragileIceFallOntoAndBreakTrigger());
 
 	private WWCriteria() {
 		throw new UnsupportedOperationException("WWCriteria contains only static declarations.");